Fix cloud function could not build
